# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Restaurant {
  id: string;
  name: string;
  shortName: string;
  address: string;
  operatorId: string;
  operatorName?: string;
  isRunning: boolean;
  /** @nullable */
  stopUntil: string | null;
  /** @nullable */
  ruleId: string | null;
  ruleActive?: boolean;
  serviceLayer?: string;
}

export type SamzaberuRequestStatus = typeof SamzaberuRequestStatus[keyof typeof SamzaberuRequestStatus];


export const SamzaberuRequestStatus = {
  REQUESTED: 'REQUESTED',
  PROCESSING: 'PROCESSING',
  COMPLETED_AUTO: 'COMPLETED_AUTO',
  ESCALATED: 'ESCALATED',
  COMPLETED_MANUAL: 'COMPLETED_MANUAL',
  CANCELLED: 'CANCELLED',
} as const;

export type SamzaberuRequestAction = typeof SamzaberuRequestAction[keyof typeof SamzaberuRequestAction];


export const SamzaberuRequestAction = {
  STOP: 'STOP',
  ENABLE: 'ENABLE',
} as const;

export interface SamzaberuRequest {
  id: string;
  status: SamzaberuRequestStatus;
  action: SamzaberuRequestAction;
  restaurantId: string;
  restaurantName: string;
  operatorId: string;
  operatorName: string;
  requestedAt: string;
  /** @nullable */
  targetUntil: string | null;
  /** @nullable */
  reason?: string | null;
  message: string;
  retryCount?: number;
  escalatedTo?: string[];
  canCompleteManually?: boolean;
}

export interface SamzaberuSummary {
  operatorId: string;
  operatorName: string;
  totalRestaurants: number;
  runningCount: number;
  stoppedCount: number;
  pendingCount: number;
  restaurants: Restaurant[];
  recentRequests: SamzaberuRequest[];
}

export type OperatorAccessAccessMode = typeof OperatorAccessAccessMode[keyof typeof OperatorAccessAccessMode];


export const OperatorAccessAccessMode = {
  full: 'full',
  assigned: 'assigned',
} as const;

export interface OperatorAccess {
  operatorId: string;
  operatorName: string;
  restaurantCount: number;
  configured: boolean;
  accessMode: OperatorAccessAccessMode;
}

export interface SamzaberuAccess {
  operators: OperatorAccess[];
}

export type SamzaberuRequestInputAction = typeof SamzaberuRequestInputAction[keyof typeof SamzaberuRequestInputAction];


export const SamzaberuRequestInputAction = {
  STOP: 'STOP',
  ENABLE: 'ENABLE',
} as const;

export interface SamzaberuRequestInput {
  action: SamzaberuRequestInputAction;
  /** @minLength 1 */
  restaurantId: string;
  /** @minLength 1 */
  operatorId: string;
  /** @nullable */
  targetUntil: string | null;
  confirmation: boolean;
  simulateFailure?: boolean;
}

export interface ManualCompletionInput {
  /** @minLength 1 */
  operatorId: string;
}

export type ListRestaurantsParams = {
operatorId?: string;
};

export type GetSamzaberuSummaryParams = {
operatorId?: string;
};

export type ListSamzaberuRequestsParams = {
operatorId?: string;
};