# lib/api-zod/src/generated/types/manualCompletionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ManualCompletionInput {
  /** @minLength 1 */
  operatorId: string;
}