# lib/api-zod/src/generated/types/operatorAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OperatorAccessAccessMode } from './operatorAccessAccessMode';

export interface OperatorAccess {
  operatorId: string;
  operatorName: string;
  restaurantCount: number;
  configured: boolean;
  accessMode: OperatorAccessAccessMode;
}