# lib/api-zod/src/generated/types/restaurant.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Restaurant {
  id: string;
  name: string;
  shortName: string;
  address: string;
  operatorId: string;
  operatorName?: string;
  isRunning: boolean;
  /** @nullable */
  stopUntil: Date | null;
  /** @nullable */
  ruleId: string | null;
  ruleActive?: boolean;
  serviceLayer?: string;
}