# lib/api-zod/src/generated/types/samzaberuAccess.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OperatorAccess } from './operatorAccess';

export interface SamzaberuAccess {
  operators: OperatorAccess[];
}