# lib/api-zod/src/generated/types/samzaberuRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SamzaberuRequestAction } from './samzaberuRequestAction';
import type { SamzaberuRequestStatus } from './samzaberuRequestStatus';

export interface SamzaberuRequest {
  id: string;
  status: SamzaberuRequestStatus;
  action: SamzaberuRequestAction;
  restaurantId: string;
  restaurantName: string;
  operatorId: string;
  operatorName: string;
  requestedAt: Date;
  /** @nullable */
  targetUntil: Date | null;
  /** @nullable */
  reason?: string | null;
  message: string;
  retryCount?: number;
  escalatedTo?: string[];
  canCompleteManually?: boolean;
}