# lib/api-zod/src/generated/types/samzaberuRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SamzaberuRequestInputAction } from './samzaberuRequestInputAction';

export interface SamzaberuRequestInput {
  action: SamzaberuRequestInputAction;
  /** @minLength 1 */
  restaurantId: string;
  /** @minLength 1 */
  operatorId: string;
  /** @nullable */
  targetUntil: Date | null;
  confirmation: boolean;
  simulateFailure?: boolean;
}