# lib/api-zod/src/generated/types/samzaberuSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Restaurant } from './restaurant';
import type { SamzaberuRequest } from './samzaberuRequest';

export interface SamzaberuSummary {
  operatorId: string;
  operatorName: string;
  totalRestaurants: number;
  runningCount: number;
  stoppedCount: number;
  pendingCount: number;
  restaurants: Restaurant[];
  recentRequests: SamzaberuRequest[];
}